# docs/swagger.yaml
basePath: /api/v1
definitions:
  AddProductSuccess:
    properties:
      data:
        example: Product added successfully.
        type: string
      success:
        example: true
        type: boolean
    type: object
  BadRequestError:
    properties:
      error:
        example: Bad Request
        type: string
      success:
        example: false
        type: boolean
    type: object
  DeleteProductSuccess:
    properties:
      data:
        example: Product successfully deleted.
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  InternalServerError:
    properties:
      error:
        example: Internal Server Error
        type: string
      success:
        example: false
        type: boolean
    type: object
...
  UnauthorizedError:
    properties:
      error:
        example: Unauthorized
        type: string
      success:
        example: false
        type: boolean
    type: object
  UpdateProductSuccess:
    properties:
      data:
        example: Product upadated successfully.
        type: string
      success:
        example: true
        type: boolean
    type: object
...
  dtos.PatchProductRequest:
    properties:
      description:
        example: Updated description
        maxLength: 500
        minLength: 2
        type: string
      imageurl:
        example: https://example.com/image.jpg
        type: string
      name:
        example: iPhone 14 Pro
        maxLength: 50
        minLength: 2
        type: string
      price:
        example: 1099.99
        type: number
      product_category_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      stock:
        example: 50
        type: integer
    type: object
  dtos.ProductRequest:
    properties:
      description:
        example: Latest Apple smartphone
        maxLength: 500
        minLength: 2
        type: string
      imageurl:
        example: https://example.com/image.jpg
        type: string
      name:
        example: iPhone 14
        maxLength: 50
        minLength: 2
        type: string
      price:
        example: 999.99
        type: number
      product_category_id:
        example: f47ac10b-58cc-4372-a567-0e02b2c3d479
        type: string
      stock:
        example: 100
        type: integer
    required:
    - name
    - price
    - product_category_id
    - stock
    type: object
...
  models.ProductResponse:
    properties:
      category:
        $ref: '#/definitions/ProductCategory'
      description:
        type: string
      imageURL:
        type: string
      name:
        type: string
      price:
        type: number
      product_id:
        type: string
      productCategoryID:
        type: string
      stock:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: jaiswal.vaibhavkumar45@gmail.com
    name: API Support
  description: API for managing products and orders.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: E-Commerce API
  version: "1.0"
paths:
...
  /product:
    get:
      consumes:
      - application/json
      description: Retrieves a list of products based on optional filters such as
        name, price, product category ID.
      parameters:
      - description: Filter by product name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by product price
        in: query
        name: price
        type: number
      - description: Filter by Product Category ID (UUID)
        in: query
        name: product_category_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of products
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Get list of products
      tags:
      - Product
    post:
      consumes:
      - application/json
      description: Creates a new product with the provided details.
      parameters:
      - description: Product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product created successfully
          schema:
            $ref: '#/definitions/AddProductSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Add a new product
      tags:
      - Product
  /product/{product_id}:
    delete:
      consumes:
      - application/json
      description: Deletes a product using its unique identifier (UUID).
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/DeleteProductSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Delete a product by ID
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Retrieves a single product by its unique identifier (UUID).
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Product details
          schema:
            $ref: '#/definitions/models.ProductResponse'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Get a product by ID
      tags:
      - Product
    patch:
      consumes:
      - application/json
      description: Partially updates an existing product using its unique ID. Only
        the provided fields will be updated.
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Fields to update in the product
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dtos.PatchProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/UpdateProductSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Partially update a product by ID
      tags:
      - Product
    put:
      consumes:
      - application/json
      description: Updates an existing product using its unique ID. All fields will
        be updated.
      parameters:
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Updated product data
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/dtos.ProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/UpdateProductSuccess'
        "400":
          description: Invalid input or missing required fields
          schema:
            $ref: '#/definitions/BadRequestError'
        "401":
          description: Unauthorized access attempt
          schema:
            $ref: '#/definitions/UnauthorizedError'
        "500":
          description: Unexpected server error
          schema:
            $ref: '#/definitions/InternalServerError'
      security:
      - BearerAuth: []
      summary: Update a product by ID
      tags:
      - Product
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"